dom/base/test/chrome/test_bug765993.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=765993
-->
<head>
  <title>Test for Bug 765993</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript"  src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=765993">Mozilla Bug 765993</a>
<style type="text/css">
#link1 a { user-select:none; }
</style>
<div id="link1"><a href="http://www.mozilla.org/">link1</a></div>
<div id="link2"><a href="http://www.mozilla.org/">link2</a></div>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 765993 **/

const {addDebuggerToGlobal} = ChromeUtils.import("resource://gre/modules/jsdebugger.jsm");
addDebuggerToGlobal(this);

window.onload = function() {
    SimpleTest.waitForExplicitFinish();

    var iframe = document.createElement("iframe");
    iframe.src = "http://mochi.test:8888/tests/dom/base/test/chrome/nochrome_bug765993.html";
    iframe.onload = function() {
        var script = iframe.contentWindow.document.createElement("script");
        script.src = "http://mochi.test:8888/tests/dom/base/test/chrome/nochrome_bug765993.js";
        script.onload = function() {
            var dbg = new Debugger(iframe.contentWindow);
            ok(dbg, "Should be able to create debugger");

            var scripts = dbg.findScripts({
                url: "http://mochi.test:8888/tests/dom/base/test/chrome/nochrome_bug765993.js",
            });
            ok(scripts.length, "Should be able to find script");

            is(scripts[0].source.sourceMapURL, "foo.js.map");
            SimpleTest.finish();
        };

        iframe.contentWindow.document.body.appendChild(script);
    };

    document.body.appendChild(iframe);
};

</script>
</pre>
</body>
</html>
